알림에서 상세페이지 바로가기 
리뷰 중복 불가
신고 기능 수정

+++ src/main/java/com/goottflix/review/model/Review.java
@@ -7,8 +7,8 @@ public class Review {
     private Long id;
     private Long movieId;
     private Long userId;
-    private Long rating;
+    private int rating;
     private String review;
-    private Long recommend;
-    private Long spoiler;
+    private int recommend;
+    private int spoiler;
 }
